api_model/requirements.txt
azureml_core==1.56.0
azureml_mlflow==1.56.0
fastapi==0.111.0
jose==1.0.0
mlflow==2.13.0
mlflow_skinny==2.13.0
pandas==2.2.2
pydantic==2.7.4
pytest==8.1.1
python-dotenv==1.0.1
python_jose==3.3.0
SQLAlchemy==2.0.29
uvicorn==0.30.1
pyodbc